Initial monthly web report implementation

+++ src/main/java/com/grupo_b/kauppa_api/product/Product.java
@@ -2,7 +2,7 @@ package com.grupo_b.kauppa_api.product;
 
 public class Product {
     private String name;
-    private Integer quantity;
+    private Long quantity;
 
     public String getName() {
         return name;
@@ -12,11 +12,11 @@ public class Product {
         this.name = name;
     }
 
-    public Integer getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Long quantity) {
         this.quantity = quantity;
     }
 }
